requirements.txt
Flask
requests
pyspark
google-cloud-bigquery
python-dotenv
gunicorn